+++ cmd/vega/genesis/load_checkpoint.go
@@ -9,16 +9,20 @@ import (
 	"os"
 	"strings"
 
+	vgfs "code.vegaprotocol.io/shared/libs/fs"
 	"code.vegaprotocol.io/vega/genesis"
 	"code.vegaprotocol.io/vega/logging"
 	vgtm "code.vegaprotocol.io/vega/tendermint"
 	"code.vegaprotocol.io/vega/types"
 	"github.com/jessevdk/go-flags"
+	tmjson "github.com/tendermint/tendermint/libs/json"
+	tmtypes "github.com/tendermint/tendermint/types"
 )
 
 type loadCheckpointCmd struct {
 	DryRun         bool   `long:"dry-run" description:"Display the genesis file without writing it"`
 	TmHome         string `short:"t" long:"tm-home" description:"The home path of tendermint"`
+	GenesisFile    string `short:"g" long:"genesis-file" description:"A genesis file to be updated"`
 	CheckpointPath string `long:"checkpoint-path" required:"true" description:"The path to the checkpoint file to load"`
 }
 
@@ -32,9 +36,18 @@ func (opts *loadCheckpointCmd) Execute(_ []string) error {
 		return err
 	}
 
-	tmConfig := vgtm.NewConfig(opts.TmHome)
+	var (
+		genesisDoc *tmtypes.GenesisDoc
+		err        error
+		tmConfig   *vgtm.Config
+	)
 
-	genesisDoc, _, err := tmConfig.Genesis()
+	if len(opts.GenesisFile) > 0 {
+		genesisDoc, _, err = readGenesisFile(opts.GenesisFile)
+	} else {
+		tmConfig = vgtm.NewConfig(opts.TmHome)
+		genesisDoc, _, err = tmConfig.Genesis()
+	}
 	if err != nil {
 		return fmt.Errorf("couldn't get genesis file: %w", err)
 	}
@@ -80,8 +93,14 @@ func (opts *loadCheckpointCmd) Execute(_ []string) error {
 	}
 
 	if !opts.DryRun {
-		if err := tmConfig.SaveGenesis(genesisDoc); err != nil {
-			return fmt.Errorf("couldn't save genesis: %w", err)
+		if len(opts.GenesisFile) > 0 {
+			if err := genesisDoc.SaveAs(opts.GenesisFile); err != nil {
+				return fmt.Errorf("couldn't save the genesis file: %w", err)
+			}
+		} else {
+			if err := tmConfig.SaveGenesis(genesisDoc); err != nil {
+				return fmt.Errorf("couldn't save genesis: %w", err)
+			}
 		}
 	}
 
@@ -91,4 +110,27 @@ func (opts *loadCheckpointCmd) Execute(_ []string) error {
 	}
 	fmt.Println(prettifiedDoc)
 	return nil
+}
+
+func readGenesisFile(path string) (*tmtypes.GenesisDoc, *genesis.GenesisState, error) {
+	data, err := vgfs.ReadFile(path)
+	if err != nil {
+		return nil, nil, fmt.Errorf("couldn't read genesis file: %w", err)
+	}
+
+	doc := &tmtypes.GenesisDoc{}
+	err = tmjson.Unmarshal(data, doc)
+	if err != nil {
+		return nil, nil, fmt.Errorf("couldn't unmarshal the genesis document: %w", err)
+	}
+
+	state := &genesis.GenesisState{}
+
+	if len(doc.AppState) != 0 {
+		if err := json.Unmarshal(doc.AppState, state); err != nil {
+			return nil, nil, fmt.Errorf("couldn't unmarshal genesis state: %w", err)
+		}
+	}
+
+	return doc, state, nil
 }